Remove the results file only when it exists.
Addressing:
Error: Could not set 'present' on ensure: No such file or directory -
/var/lib/openscap/xccdf_scan/untitled/2014-07-19.rds.xml.bz2 at
72:/etc/puppet/environments/production/modules/openscap/manifests/xccdf/foreman_audit.pp
Wrapped exception:
No such file or directory -
/var/lib/openscap/xccdf_scan/untitled/2014-07-19.rds.xml.bz2
Error:
/Stage[main]/Openscap::Xccdf::Foreman_audit/Xccdf_scan[untitled]/ensure:
change from absent to present failed: Could not set 'present' on ensure:
No such file or directory -
/var/lib/openscap/xccdf_scan/untitled/2014-07-19.rds.xml.bz2 at
72:/etc/puppet/environments/production/modules/openscap/manifests/xccdf/foreman_audit.pp

+++ lib/puppet/provider/xccdf_scan/openscap.rb
@@ -37,7 +37,7 @@ Puppet::Type.type(:xccdf_scan).provide :openscap do
       bzip2 _target_location_rds
       scap_upload.provider.upload self if scap_upload
     rescue Exception => e
-      File.unlink result_path
+      File.unlink result_path if exists?
       raise
     end
   end
